test: consolidate remaining unit tests under test/unit
Relocate the last co-located unit tests (auth, ssrf, and the metrics/pdf/
screenshot/storage services) from src into test/unit so there is a single
home for unit tests, alongside test/integration. Relative imports and the
relative module mocks (env in StorageService and auth) updated to resolve
from the new location.
